Implement Mantis Lords (#39)
* Implement Mantis Lords (WIP)

The animations are mostly synced. There are still several issues though:
- These entities spawn in when challenged, not when the scene changes. This causes the initializers to only be called when challenging the Mantis Lords yourself.
- When the second phase starts, the initializer is thus not called for the clients and this means that I can't disable the battle logic of the second phase (which is not in the Mantis Lord FSM)
- The downward stab move's hitbox is not correct, and sometimes hits a client from far away (don't know what causes this)

* Implement Mantis Lord second phase and challenging

Most of the fight works now, except for a few things:
- The Music animation gets send on the host side, but does not get received
- Some issues with switching between host and client
- Did not test in Godhome yet

* Fix Mantis lords in Godhome and when switching between host and client

Whether we're in Godhome is checked by seeing if there is a challenge object, which is only in Mantis Village.
Also, there were a bunch of bugs where joining the fight late and switching between host and client would stop or crash the fight.

* Apply suggestions from code review

Co-authored-by: Extremelyd1 <Extremelyd1@outlook.com>

* Fix PR suggestions and add a way for the entity to for end the fight

I found that Lord S1 and S2 are not actually bound to throne 1 and 3.
If one dies, throne 3 will always have the defeated state.
This only causes issues when someone joins the fight when one of mantis lords is dead in the second phase. That's why when the host ends the fight, it forces the fight to end on the client side.

* Move ternary operator to an if statement

Co-authored-by: Extremelyd1 <Extremelyd1@outlook.com>